requirements.txt
Click==8.1.7
htmlmin==0.1.12
Jinja2==3.1.3
jsmin==3.0.1
livereload==2.6.3
# mkdocs 2.4.1 requires Markdown < 3.4.0
# https://github.com/kubernetes-sigs/gateway-api/pull/1671#issuecomment-1400586465
markdown~=3.6
MarkupSafe==2.1.5
mkdocs==1.6.0
mkdocs-awesome-pages-plugin==2.9.2
mkdocs-macros-plugin==1.0.5
mkdocs-material==9.5.19
mkdocs-material-extensions==1.3.1
mkdocs-redirects==1.2.1
mkdocs-mermaid2-plugin==1.1.1
pandas==2.0.3
pep562==1.1
Pygments==2.17.2
pymdown-extensions==10.8.1
PyYAML==6.0.1
six==1.16.0
tabulate==0.9.0
tornado==6.4
